# By Sebastian Freundt (30) and Ruediger Meier (1)
* next: (31 commits)
  fix, pass MREMAP_MAYMOVE to all mremap() calls
  test, hack, hard-code the size of the reference file in fsck.22.ut regression
  test, chore, use good old cp without any args for fsck.24.ut regression
  fix, PEBKAC, initialise IDX local variable in ut-test.c
  chore, define _XOPEN_SOURCE for C files in test/ as well
  chore, define _XOPEN_SOURCE, wethinks 6.00 should be good enough
  chore, #undef pthread_mutex_*() macros before redefining them
  test, hygiene, adhere to POSIX' mktemp calling convention (i.e. use a template)
  test, hygiene, use boring old cp
  test, hygiene, adhere to POSIX' mktemp calling convention (i.e. use a template)
  test, chore, avoid the use of $(..) expansion in all regression test files
  test, cosmetics, kick big_endian_p() definition in mux.4.ut
  chore, don't distribute binary ut-test (log compiler) but the shell part of it
  chore, distribute built sources
  minor, test, put init'ing effort of ut-test.sh into ut-test binary (can also cope with longopts)
  chore, free test suite script from $(...) clauses and use portable backticks instead
  fix, PEBKAC, make sure memory maps are always against file descriptor -1
  chore, include sys/stat.h for stat operations in ute-fsck
  fix, PEBKAC, it's htooe16() not __htooe16()
  fix, unsure, when htooe* macros can't be defined at least make
  ...

+++ src/ute-print.h
@@ -38,6 +38,8 @@
 #define INCLUDED_ute_print_h_
 
 #include <string.h>
+/* for ssize_t */
+#include <unistd.h>
 #include "scommon.h"
 #include "nifty.h"
 
@@ -146,8 +148,10 @@ print_tick_sym(pr_ctx_t pctx, scom_t st)
 
 /**
  * Public print function.
- * Implemented through DSOs. */
-extern ssize_t pr(pr_ctx_t pctx, scom_t st);
+ * Implemented through DSOs.
+ * The return value should indicate whether or not the tick in ST
+ * has been processed, a value <0 indicates failure. */
+extern int pr(pr_ctx_t pctx, scom_t st);
 
 /**
  * For printers that need initialisation and finalisation. */
